Expense Tracker/requirements.txt
asgiref==3.8.1
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2024.1
Django==5.0.3
django-cors-headers==4.3.1
django-filter==24.2
django-taggit==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
PyJWT==2.8.0
requests==2.31.0

